requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
torchmetrics>=1.0.0

# Data processing
numpy>=1.24.0
pillow>=9.5.0
opencv-python>=4.8.0
albumentations>=1.3.0

# Visualization
matplotlib>=3.7.0
tensorboard>=2.13.0

# Configuration
pyyaml>=6.0

# Optional but recommended
tqdm>=4.65.0
scikit-learn>=1.3.0
mpl_toolkits  # For 3D plotting
pandas>=2.0.0